adds the WhatsApp new-conversation UI
- New conversation dialog now shows an Email / WhatsApp tab toggle, but only when a WhatsApp inbox exists. With no WhatsApp inbox it stays the plain email form, no tabs.
- The email form is extracted into EmailConversationForm.vue unchanged. WhatsAppConversationForm.vue is new: search/select a contact (or enter a new number with the country picker), pick an approved template, fill its params, and send.
- Template picking logic is shared via useWhatsAppTemplatePicker.js and a presentational WhatsAppTemplatePicker.vue, used by both the new form and the existing reply composer. The composer was refactored to use them.
- The contact search dropdown shows name, phone, email and external id so the agent can confirm the right person. The selected chip shows just name and phone.
- inbox store gains whatsappOptions, mirroring emailOptions.
