+++ frontend/apps/main/src/stores/inbox.js
@@ -16,6 +16,10 @@ export const useInboxStore = defineStore('inbox', () => {
     .filter(inb => inb.channel === 'email')
     .map(inb => ({ label: inb.name, value: String(inb.id) }))
   )
+  const whatsappOptions = computed(() => inboxes.value
+    .filter(inb => inb.channel === 'whatsapp')
+    .map(inb => ({ label: inb.name, value: String(inb.id) }))
+  )
   const fetchInboxes = async (force = false) => {
     if (!force && inboxes.value.length) return
     try {
@@ -32,6 +36,7 @@ export const useInboxStore = defineStore('inbox', () => {
     inboxes,
     options,
     emailOptions,
+    whatsappOptions,
     fetchInboxes,
   }
 })
